logging: Extract from ffi.rs to log.rs.
